FE/Qt4: warnings
svn:sync-xref-src-repo-rev: r68313

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIKeyboardHandler.cpp
@@ -1,4 +1,4 @@
-/* $Id: UIKeyboardHandler.cpp 34476 2010-11-29 15:44:39Z sergey.dubov@oracle.com $ */
+/* $Id: UIKeyboardHandler.cpp 34523 2010-11-30 15:17:06Z noreply@oracle.com $ */
 /** @file
  *
  * VBox frontends: Qt GUI ("VirtualBox"):
@@ -135,7 +135,7 @@ void UIKeyboardHandler::prepareListener(ulong uIndex, UIMachineWindow *pMachineW
 void UIKeyboardHandler::cleanupListener(ulong uIndex)
 {
     /* Check if we should release keyboard first: */
-    if (uIndex == m_iKeyboardCaptureViewIndex)
+    if ((int)uIndex == m_iKeyboardCaptureViewIndex)
         releaseKeyboard();
 
     /* If window still registered: */
